requirements.txt
fastapi==0.115.*
uvicorn[standard]==0.32.*
asyncpg==0.30.*
pydantic-settings==2.*
python-multipart==0.0.*
bcrypt==4.*
PyJWT==2.*
httpx==0.27.*
reportlab==4.*
python-dateutil==2.*
pytest==8.*
